TECHNIQUE/03_C_MOTEURS_ET_DEPLOIEMENT/ynor_llm_gateway.py
# **[◬] MATRICE FRACTALE MDL YNOR V2.0**
# **Corpus :** MDL YNOR
# **Position Structurelle :** LAYER
# **Position Chiastique :** C
# **Rôle du Fichier :** Script ou configuration declarative
# **Centre Doctrinal Local :** AI Manager garde script ou configuration declarative en limitant le bruit local et la friction structurelle.
# **Loi de Survie :** μ = α - β - κ
# **Lecture Locale :**
# - **α :** stabilite locale
# - **β :** bruit externe injecte
# - **κ :** friction structurelle
# **Risque :** e∞ ∝ ε / μ
# **Opérateur Correctif :** D(S)=proj_{SafeDomain}(S)
# **Axiome :** un système survit SSI μ > 0
# **Doctrine Goodhart : tout succès apparent est invalide si μ décroît**
# **Gouvernance : toute modification doit maximiser Δμ**
# **Lien Miroir :** C'

# SDKs
from openai import AsyncOpenAI
import google.generativeai as genai

# =================================================================================
# MDL YNOR - LLM GATEWAY V2.0 [REAL-WORLD INTEGRATION]
# Principe d'Invariance : mu = Intersection(Actions_Safe_Models)
# =================================================================================

load_dotenv()

class YnorLLMGateway:
    """
    Passerelle unifiée pour l'audit multi-modèles (Problème #4: Intégration).
    Connecte le moteur Ynor aux APIs industrielles pour validation empirique.
    """
    def __init__(self):
        # Initialisation OpenAI
        self.openai_client = AsyncOpenAI(api_key=os.getenv("OPENAI_API_KEY"))
        
        # Initialisation Google Gemini
        genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))
        
        self.providers = {
            "openai": ["gpt-4o", "gpt-4o-mini", "o1-preview"],
            "google": ["gemini-1.5-pro", "gemini-1.5-flash", "gemini-2.0-flash-exp"],
            "anthropic": ["claude-3-5-sonnet-latest"]
        }

    async def call_model(self, provider: str, model: str, prompt: str, temperature: float = 0.0) -> Dict[str, Any]:
        """
        Appel asynchrone réel aux APIs (Problème #1: Validation Empirique).
        """
        try:
            if provider == "openai":
                response = await self.openai_client.chat.completions.create(
                    model=model,
                    messages=[{"role": "user", "content": prompt}],
                    temperature=temperature
                )
                return {
                    "text": response.choices[0].message.content,
                    "usage": response.usage.total_tokens,
                    "model": model,
                    "status": "success"
                }

            elif provider == "google":
                gemini_model = genai.GenerativeModel(model)
                # GenerationConfig pour forcer la stabilité (temperature 0)
                response = await gemini_model.generate_content_async(
                    prompt, 
                    generation_config=genai.types.GenerationConfig(temperature=temperature)
                )
                return {
                    "text": response.text,
                    "usage": 0, # Placeholder pour Gemini usage
                    "model": model,
                    "status": "success"
                }
            
            # Placeholder pour Anthropic
            return {"error": f"Provider {provider} implementation pending", "status": "pending"}

        except Exception as e:
            return {"error": str(e), "model": model, "status": "failed", "text": ""}

    async def cross_model_audit(self, prompt: str, selected_models: List[Dict[str, str]]) -> Dict[str, Any]:
        """
        Exécute le même prompt sur plusieurs modèles en parallèle pour mesurer la dérive (Problème #3: Mesure mu).
        """
        tasks = []
        for m in selected_models:
            tasks.append(self.call_model(m["provider"], m["model"], prompt))
        
        results = await asyncio.gather(*tasks)
        
        # Nettoyage des résultats échoués
        valid_responses = [r for r in results if r["status"] == "success"]
        
        return {
            "responses": valid_responses,
            "count": len(valid_responses),
            "timestamp": os.getenv("TIMESTAMP_CANONICAL", "2026-04-12")
        }

if __name__ == "__main__":
    # Test Rapide (nécessite des clés API dans .env)
    async def main():
        gateway = YnorLLMGateway()
        print("--- YNOR GATEWAY V2.0 : SYSTEM READY ---")
        # test_res = await gateway.call_model("google", "gemini-1.5-flash", "What is MDL Ynor?")
        # print(test_res)

    asyncio.run(main())
