TECHNIQUE/03_C_MOTEURS_ET_DEPLOIEMENT/requirements.txt
fastapi

uvicorn[standard]

pydantic<2

openai

anthropic

google-generativeai

python-dotenv

numpy

requests

langchain

langchain-openai

langchain-anthropic

langchain-google-genai

langgraph

pandas

yfinance

stockstats

gunicorn

scipy

slowapi

PyJWT
